refactor: move InvestmentScope enum out of Interfaces/ into new Enums/ folder
InvestmentScope is a plain enum, not an interface - it was sitting in
Financial.Application/Interfaces/ purely because that's where it was
first added, not because it belongs there. Found during an audit of
whether Application layer interfaces are correctly placed and used.

Moved to Financial.Application/Enums/InvestmentScope.cs (namespace
Financial.Application.Enums). Mechanical follow-through: every consumer
across Interfaces, Services, Validation, Infrastructure, the WPF app,
and both test projects needed the new using added (kept alongside the
existing Interfaces using wherever that file also depends on IRepository/
INavigationService/etc.; replaced it where InvestmentScope was the only
reason for the import).
