+++ Financial.App/ViewModels/MainNavigationViewModelBase.cs
@@ -1,4 +1,5 @@
 using Financial.Application.DTOs;
+using Financial.Application.Enums;
 using Financial.Application.Interfaces;
 using Financial.Domain.Entities;
 using System.Collections.ObjectModel;
