+++ Financial.Application/Services/NavigationMapper.cs
@@ -1,5 +1,5 @@
 using Financial.Application.DTOs;
-using Financial.Application.Interfaces;
+using Financial.Application.Enums;
 using Financial.Domain.Entities;
 
 namespace Financial.Application.Services;
