+++ Tests/Financial.Application.Tests/Services/TransactionServiceQueryTests.cs
@@ -1,3 +1,4 @@
+using Financial.Application.Enums;
 using Financial.Application.Interfaces;
 using Financial.Application.Services;
 using Financial.Domain.Entities;
