+++ Tests/Financial.Presentation.Tests/ViewModels/AssetDetailsViewModelXirrTests.cs
@@ -1,4 +1,5 @@
 using Financial.Application.DTOs;
+using Financial.Application.Enums;
 using Financial.Application.Interfaces;
 using Financial.Application.Services;
 using Financial.Domain.Entities;
